+++ quantumphysique/lib/quantumphysique.dart
@@ -55,6 +55,8 @@ export 'src/widgets/sliver_app_bar_snap.dart';
 export 'src/widgets/selection_carousel.dart';
 export 'src/widgets/third_party_licence.dart';
 export 'src/widgets/qp_m3e_fab.dart';
+export 'src/widgets/qp_filter_chip.dart';
+export 'src/widgets/qp_button_group.dart';
 
 // Pages
 export 'src/pages/qp_home_page.dart';